Make babel and its plugins regular dependencies
This needs to be like this way for now, since we're building supernova after installing it (which is an anti pattern... we should instead already sent the built files to whoever installs supernova...). Otherwise,  people installing supernova also have to install the others manually (and figure out that they need to do that as well, which is not so straightforward).